+++ apps/tests/aws-runtime/test/test-service.ts
@@ -1,3 +1,4 @@
+import z from "zod";
 import {
   activity,
   condition,
@@ -312,26 +313,30 @@ export const eventDrivenWorkflow = workflow(
   }
 );
 
-const signalEvent = event<{
-  executionId: string;
-  signalId: string;
-  proxy?: true;
-}>("SignalEvent");
-
-signalEvent.onEvent(async ({ executionId, signalId, proxy }) => {
-  console.debug("received signal event", { executionId, signalId, proxy });
-  if (proxy) {
-    // if configured to proxy, re-route this event through the signalEvent
-    // reason: to test that we can publish events from within an event handler
-    await signalEvent.publishEvents({
-      executionId,
-      signalId,
-    });
-  } else {
-    // otherwise, send the signal to the workflow
-    await sendSignal(executionId, signalId, { value: "done!" });
-  }
+const SignalEventPayload = z.object({
+  executionId: z.string(),
+  signalId: z.string(),
+  proxy: z.literal(true).optional(),
 });
+
+const signalEvent = event("SignalEvent", SignalEventPayload);
+
+export const onSignalEvent = signalEvent.onEvent(
+  async ({ executionId, signalId, proxy }) => {
+    console.debug("received signal event", { executionId, signalId, proxy });
+    if (proxy) {
+      // if configured to proxy, re-route this event through the signalEvent
+      // reason: to test that we can publish events from within an event handler
+      await signalEvent.publishEvents({
+        executionId,
+        signalId,
+      });
+    } else {
+      // otherwise, send the signal to the workflow
+      await sendSignal(executionId, signalId, { value: "done!" });
+    }
+  }
+);
 
 const sendFinishEvent = activity("sendFinish", async (executionId: string) => {
   // publish an event from an activity