+++ packages/@eventual/aws-cdk/src/service.ts
@@ -37,6 +37,7 @@ import path from "path";
 import { Activities, IActivities } from "./activities";
 import { BuildOutput, buildServiceSync } from "./build";
 import { Events } from "./events";
+import { grant } from "./grant";
 import { Logging, LoggingProps } from "./logging";
 import { lazyInterface } from "./proxy-construct";
 import { IScheduler, Scheduler } from "./scheduler";
@@ -385,7 +386,10 @@ export class Service extends Construct implements IGrantable, IService {
   public addEnvironment(key: string, value: string): void {
     this.activities.worker.addEnvironment(key, value);
     this.api.handlers.forEach((handler) => handler.addEnvironment(key, value));
-    this.events.handler.addEnvironment(key, value);
+    this.events.defaultHandler.addEnvironment(key, value);
+    this.events.handlers.forEach((handler) =>
+      handler.addEnvironment(key, value)
+    );
     this.workflows.orchestrator.addEnvironment(key, value);
   }
 
@@ -397,6 +401,7 @@ export class Service extends Construct implements IGrantable, IService {
     this.workflows.configureStartExecution(func);
   }
 
+  @grant()
   public grantStartExecution(grantable: IGrantable) {
     this.workflows.grantStartExecution(grantable);
   }
@@ -406,7 +411,7 @@ export class Service extends Construct implements IGrantable, IService {
     this.workflows.configureReadExecutionHistory(func);
     this.workflows.configureReadHistoryState(func);
   }
-
+  @grant()
   public grantReadExecutions(grantable: IGrantable) {
     this.workflows.grantReadExecutions(grantable);
   }
@@ -415,6 +420,7 @@ export class Service extends Construct implements IGrantable, IService {
     this.workflows.configureSendSignal(func);
   }
 
+  @grant()
   public grantSendSignal(grantable: IGrantable) {
     this.workflows.grantSendSignal(grantable);
   }
@@ -423,6 +429,7 @@ export class Service extends Construct implements IGrantable, IService {
     this.events.configurePublish(func);
   }
 
+  @grant()
   public grantPublishEvents(grantable: IGrantable) {
     this.events.grantPublish(grantable);
   }