+++ packages/@eventual/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./activity.js";
+export * from "./api-request.js";
 export * from "./api.js";
 export * from "./await-all-settled.js";
 export * from "./await-all.js";
@@ -17,6 +18,7 @@ export * from "./execution.js";
 export * from "./flags.js";
 export * from "./global.js";
 export * from "./heartbeat.js";
+export * from "./http-method.js";
 export * from "./logging.js";
 export * from "./race.js";
 export * from "./result.js";