+++ packages/@eventual/core/src/service-spec.ts
@@ -1,20 +1,38 @@
-import type { HttpMethod } from "./api.js";
-import type { Subscription } from "./event.js";
+import type { EventHandler, Subscription } from "./event.js";
+import type { HttpMethod } from "./http-method.js";
 import type { DurationSchedule } from "./schedule.js";
+import type { SchemaObject } from "openapi3-ts";
 
 /**
  * Specification for an Eventual application
  */
 export interface ServiceSpec {
-  /**
-   * A list of all event {@link Subscription}s.
-   */
-  subscriptions: Subscription[];
   api: ApiSpec;
+  events: EventSpec;
   /**
    * List of workflows
    */
   workflows: WorkflowSpec[];
+}
+
+export interface EventSpec {
+  /**
+   * Schemas of all events within this Service.
+   */
+  schemas: Schemas;
+  /**
+   * Catch-all default subscriptions and route to the default Event Handler monolith.
+   */
+  subscriptions: Subscription[];
+  /**
+   * Individually bundled and subscribed event Event Handlers.
+   */
+  handlers: EventHandlerSpec[];
+}
+
+export interface EventHandlerSpec extends Omit<EventHandler, "handler"> {
+  // source location is mandatory for individually bundled event handlers.
+  sourceLocation: SourceLocation;
 }
 
 export interface ApiSpec {
@@ -32,9 +50,22 @@ export interface RouteSpec extends FunctionSpec {
   sourceLocation?: SourceLocation;
 }
 
+export function isSourceLocation(a: any) {
+  return (
+    a &&
+    typeof a === "object" &&
+    typeof a.fileName === "string" &&
+    typeof a.exportName === "string"
+  );
+}
+
 export interface SourceLocation {
-  fileName?: string;
-  exportName?: string;
+  fileName: string;
+  exportName: string;
+}
+
+export interface Schemas {
+  [schemaName: string]: SchemaObject;
 }
 
 export interface WorkflowSpec {